# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "PaidPayableItemResponse": {
        "description": "Conta paga incluída no relatório.",
        "properties": {
          "amount": {
            "description": "Valor pago.",
            "example": 1500.0,
            "type": "number"
          },
          "description": {
            "description": "Descrição da conta.",
            "example": "Pagamento de aluguel",
            "type": "string"
          },
          "expirationDate": {
            "description": "Data de vencimento.",
            "example": "2026-08-10",
            "format": "date",
            "type": "string"
          },
          "paymentDate": {
            "description": "Data do pagamento.",
            "example": "2026-08-05",
            "format": "date",
            "type": "string"
          },
          "status": {
            "description": "Status da conta.",
            "enum": [
              "PAGO"
            ],
            "example": "PAGO",
            "type": "string"
          },
          "supplierName": {
            "description": "Nome do fornecedor.",
            "example": "TOTVS",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "Erro HTTP no formato RFC 9457.",
        "properties": {
          "detail": {
            "description": "Explicação específica do erro.",
            "example": "Dados inválidos",
            "type": "string"
          },
          "instance": {
            "description": "URI da requisição que originou o erro.",
            "type": "string"
          },
          "status": {
            "description": "Código HTTP.",
            "example": 400,
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "description": "Título correspondente ao status HTTP.",
            "example": "Bad Request",
            "type": "string"
          },
          "type": {
            "description": "Identificador do tipo do problema.",
            "example": "about:blank",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TotalPaidReportResponse": {
        "description": "Consolidação das contas pagas no período.",
        "properties": {
          "endDate": {
            "description": "Fim do período.",
            "example": "2026-08-31",
            "format": "date",
            "type": "string"
          },
          "paidCount": {
            "description": "Quantidade de contas pagas.",
            "example": 2,
            "format": "int32",
            "type": "integer"
          },
          "paidPayableItems": {
            "description": "Contas consideradas no cálculo.",
            "items": {
              "$ref": "#/components/schemas/PaidPayableItemResponse"
            },
            "type": "array"
          },
          "startDate": {
            "description": "Início do período.",
            "example": "2026-08-01",
            "format": "date",
            "type": "string"
          },
          "totalPaid": {
            "description": "Soma dos valores pagos.",
            "example": 2390.9,
            "type": "number"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "description": "Credenciais configuradas por APP_SECURITY_USERNAME e APP_SECURITY_PASSWORD.",
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para gestão de fornecedores, contas a pagar, importações assíncronas em CSV\ne relatórios de pagamentos. Todos os recursos exigem autenticação HTTP Basic.\n",
    "title": "Payable Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/report/total-paid": {
      "get": {
        "description": "Consolida as contas com status PAGO cuja paymentDate esteja no período inclusivo.\nsupplierId é opcional; startDate e endDate são obrigatórias e devem formar um\nperíodo válido.\n",
        "operationId": "generateTotalPaidReport",
        "parameters": [
          {
            "description": "Fornecedor opcional para restringir o relatório.",
            "example": 1,
            "in": "query",
            "name": "supplierId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Data inicial inclusiva no formato ISO-8601.",
            "example": "2026-08-01",
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Data final inclusiva no formato ISO-8601.",
            "example": "2026-08-31",
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalPaidReportResponse"
                }
              }
            },
            "description": "Relatório consolidado."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Período ausente ou inválido."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Fornecedor não existe ou não há pagamentos no período."
          }
        },
        "summary": "Gerar relatório de pagamentos",
        "tags": [
          "Relatórios"
        ]
      }
    },
...
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "servers": [
    {
      "description": "Servidor da aplicação",
      "url": "/management/rest"
    }
  ],
...
}